MatlabMontageReader now unpacks nested structs; created separate file for electrode-related tests
