+++ cmlreaders/constants.py
@@ -61,8 +61,14 @@ rhino_paths = {
     'pairs': [
         'protocols/{protocol}/subjects/{subject}/localizations/{localization}/montages/{montage}/neuroradiology/current_processed/pairs.json',
     ],
+    'matlab_pairs': [
+        'data/eeg/{subject_montage}/tal/{subject_montage}_talLocs_database_bipol.mat',
+    ],
     'contacts': [
         'protocols/{protocol}/subjects/{subject}/localizations/{localization}/montages/{montage}/neuroradiology/current_processed/contacts.json',
+    ],
+    'matlab_contacts': [
+        'data/eeg/{subject_montage}/tal/{subject_montage}_talLocs_database_monopol.mat',
     ],
 
     # Report Data
